Refaktoroi be:n adaptereita ja siirrä funkkareita omiin tiedostoihinsa.
