+++ backend/src/handler/adaptProjektiUtil/adaptSuunitteluSopimus.ts
@@ -1,0 +1,17 @@
+import { SuunnitteluSopimus } from "../../database/model";
+import * as API from "../../../../common/graphql/apiModel";
+import { fileService } from "../../files/fileService";
+
+export function adaptSuunnitteluSopimus(
+  oid: string,
+  suunnitteluSopimus?: SuunnitteluSopimus | null
+): API.SuunnitteluSopimus | undefined | null {
+  if (suunnitteluSopimus) {
+    return {
+      __typename: "SuunnitteluSopimus",
+      ...suunnitteluSopimus,
+      logo: fileService.getYllapitoPathForProjektiFile(oid, suunnitteluSopimus.logo),
+    };
+  }
+  return suunnitteluSopimus as undefined | null;
+}
