+++ backend/src/handler/adaptProjektiUtil/common.ts
@@ -1,0 +1,17 @@
+import * as API from "../../../../common/graphql/apiModel";
+
+export function adaptIlmoituksenVastaanottajat(
+  vastaanottajat: API.IlmoituksenVastaanottajat | null | undefined
+): API.IlmoituksenVastaanottajat {
+  if (!vastaanottajat) {
+    return vastaanottajat as null | undefined;
+  }
+  const kunnat: API.KuntaVastaanottaja[] =
+    vastaanottajat?.kunnat?.map((kunta) => ({ __typename: "KuntaVastaanottaja", ...kunta })) || null;
+  const viranomaiset: API.ViranomaisVastaanottaja[] =
+    vastaanottajat?.viranomaiset?.map((viranomainen) => ({
+      __typename: "ViranomaisVastaanottaja",
+      ...viranomainen,
+    })) || null;
+  return { __typename: "IlmoituksenVastaanottajat", kunnat, viranomaiset };
+}
