+++ backend/src/handler/commonAdapterUtil/lisaaTypename.ts
@@ -1,0 +1,50 @@
+import { Kielitiedot, Linkki, Suunnitelma, Velho } from "../../database/model";
+import * as API from "../../../../common/graphql/apiModel";
+
+export function adaptLiittyvatSuunnitelmat(suunnitelmat?: Suunnitelma[] | null): API.Suunnitelma[] | undefined | null {
+  if (suunnitelmat) {
+    const liittyvatSuunnitelmat = suunnitelmat.map(
+      (suunnitelma) =>
+        ({
+          __typename: "Suunnitelma",
+          ...suunnitelma,
+        } as Suunnitelma)
+    );
+    return liittyvatSuunnitelmat as API.Suunnitelma[];
+  }
+  return suunnitelmat as undefined | null;
+}
+
+export function adaptKielitiedot(kielitiedot?: Kielitiedot | null): API.Kielitiedot | undefined | null {
+  if (kielitiedot) {
+    return {
+      ...kielitiedot,
+      __typename: "Kielitiedot",
+    };
+  }
+  return kielitiedot as undefined;
+}
+
+export function adaptLinkki(link: Linkki): API.Linkki {
+  if (link) {
+    return {
+      ...link,
+      __typename: "Linkki",
+    };
+  }
+  return link as undefined;
+}
+
+export function adaptLinkkiList(links: Array<Linkki>): API.Linkki[] {
+  if (links) {
+    return links.map((link) => ({
+      ...link,
+      __typename: "Linkki",
+    }));
+  }
+  return links as undefined;
+}
+
+export function adaptVelho(velho: Velho): API.Velho {
+  return { __typename: "Velho", ...velho };
+}
